+++ typesafe_firebase_functions/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.3
 
 * More options for https callable functions.
+* Sure httpscallable stream.
 
 ## 0.0.2